% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_console_monitor.r
\name{app_log}
\alias{app_log}
\title{Helper to append log}
\usage{
app_log(msg, type = "info")
}
\arguments{
\item{msg}{Message to log.}

\item{type}{Log type, such as `"info"`, `"warning"`, or `"error"`.}
}
\description{
Call this function from anywhere in your app to log to the console widget.
}
