% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_feature_network.R
\name{as_feature_igraph}
\alias{as_feature_igraph}
\title{Convert a feature network edge table to an igraph object}
\usage{
as_feature_igraph(object, feature_network = NULL)
}
\arguments{
\item{object}{A `mass_dataset` object.}

\item{feature_network}{Optional edge table. If `NULL`, the network stored in
`object` is used.}
}
\value{
An `igraph` object.
}
\description{
Convert a feature network edge table to an igraph object
}
