% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_outlier.R
\name{calc_outlier_table}
\alias{calc_outlier_table}
\title{Calculate Outlier Table (Simulating masscleaner::extract_outlier_table structure)}
\usage{
calc_outlier_table(object, th_na, th_sd, th_mad, th_dist)
}
\arguments{
\item{object}{A mass_dataset object}

\item{th_na}{NA percentage threshold (0-1)}

\item{th_sd}{SD Fold Change threshold}

\item{th_mad}{MAD Fold Change threshold}

\item{th_dist}{PCA Mahalanobis distance P-value threshold}
}
\value{
A data.frame with logical columns: according_to_na, pc_sd, pc_mad, according_to_distance
}
\description{
Calculate Outlier Table (Simulating masscleaner::extract_outlier_table structure)
}
