% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_feature_network.R
\name{default_plant_neutral_loss_table}
\alias{default_plant_neutral_loss_table}
\alias{default_plant_fragment_ion_table}
\title{Plant-focused neutral loss and fragment ion dictionaries}
\usage{
default_plant_neutral_loss_table()

default_plant_fragment_ion_table()
}
\value{
A data frame.
}
\description{
Plant-focused neutral loss and fragment ion dictionaries
}
