% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_qc_report.R
\name{export_qc_report}
\alias{export_qc_report}
\title{Export QC Report (Quarto)}
\usage{
export_qc_report(object_pos, object_neg, plot_params, file)
}
\arguments{
\item{object_pos}{Positive mode mass_dataset object (or NULL)}

\item{object_neg}{Negative mode mass_dataset object (or NULL)}

\item{plot_params}{A list containing all plotting parameters}

\item{file}{Output file path provided by downloadHandler}
}
\description{
Generates an HTML report using the parameterized Quarto template.
}
