% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_overview.R
\name{mod_data_overview_server}
\alias{mod_data_overview_server}
\title{Data Overview Server Module}
\usage{
mod_data_overview_server(id, global_data, downloads)
}
\description{
Data Overview Server Module
}
