% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_project_init.R
\name{summarize_massdataset}
\alias{summarize_massdataset}
\title{Summarize mass_dataset Object Information}
\usage{
summarize_massdataset(
  object,
  mode = c("positive", "negative"),
  show_processing = TRUE,
  show_qc = TRUE,
  color = TRUE
)
}
\arguments{
\item{object}{A mass_dataset object}

\item{mode}{Ionization mode}

\item{show_processing}{Logical}

\item{show_qc}{Logical}

\item{color}{Logical}
}
\description{
Generates a formatted report summarizing key metadata.
}
