% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_project_init.R
\name{validate_file}
\alias{validate_file}
\title{Validate mass_dataset file}
\usage{
validate_file(path, expected_polarity, object_label)
}
\arguments{
\item{path}{Path to an `.rda` file containing a `mass_dataset` object.}

\item{expected_polarity}{Expected polarity label.}

\item{object_label}{Human-readable label used in error messages.}
}
\description{
Validate mass_dataset file
}
