% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_import.R
\name{validate_sample_files}
\alias{validate_sample_files}
\title{Validate sample files match sample info}
\usage{
validate_sample_files(
  mode = "POS",
  QC_num.p,
  S_num.p,
  QC_num.n,
  S_num.n,
  sample_info_temp
)
}
\arguments{
\item{mode}{"POS" or "NEG"}

\item{QC_num.p}{List of QC files (POS)}

\item{S_num.p}{List of Subject files (POS)}

\item{QC_num.n}{List of QC files (NEG)}

\item{S_num.n}{List of Subject files (NEG)}

\item{sample_info_temp}{Sample info dataframe}
}
\description{
Validate sample files match sample info
}
