% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_console_monitor.r
\name{app_log}
\alias{app_log}
\title{Helper to append log
Call this function from anywhere in your app to log to the console widget.}
\usage{
app_log(msg, type = "info")
}
\description{
Helper to append log
Call this function from anywhere in your app to log to the console widget.
}
