% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_project_init.R
\name{check_massdata_info}
\alias{check_massdata_info}
\title{Generate Formatted Data Summary Report for Shiny}
\usage{
check_massdata_info(object, mode = c("positive", "negative"))
}
\arguments{
\item{object}{A `mass_dataset` object or a reactive returning one.}

\item{mode}{Ionization mode.}
}
\description{
Generate Formatted Data Summary Report for Shiny
}
