% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_import.R
\name{check_ms1_structure}
\alias{check_ms1_structure}
\title{Helper to check directory structure for Raw Import}
\usage{
check_ms1_structure(path)
}
\description{
Helper to check directory structure for Raw Import
}
