% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_rm_noise.R
\name{find_noise_multiple}
\alias{find_noise_multiple}
\title{Find and Remove Noise}
\usage{
find_noise_multiple(object, tag = "class", qc_na_freq = 0.2, S_na_freq = 0.5)
}
\arguments{
\item{object}{a mass_dataset class}

\item{tag}{Remove noise based on column names of `sample_info` in object}

\item{qc_na_freq}{missing value ratio of QC sample (threshold)}

\item{S_na_freq}{missing value ratio of tag groups (threshold)}
}
\value{
A list containing `object_mv` (cleaned object) and `noisy_tbl` (removed features info).
}
\description{
Detected errors or batch effects of your untargeted metabolomics data.
Based on missing value frequency in QC and Sample groups.
}
