% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_paramounter.R
\name{paramounter_part1}
\alias{paramounter_part1}
\title{Paramounter part1 (Optimized)}
\usage{
paramounter_part1(
  directory,
  massSDrange = 2,
  smooth = 0,
  cutoff = 0.95,
  thread = 1,
  filenum = c(3, 5, "all")
)
}
\arguments{
\item{directory}{The data path for mzXML files.}

\item{massSDrange}{The range of standard deviations for mass differences, default 2.}

\item{smooth}{Chromatographic smoothness level (default 0).}

\item{cutoff}{The ppm percentage for cutoff (default 0.95).}

\item{thread}{Number of threads for parallel processing.}

\item{filenum}{Number of files to use (3, 5, or "all").}
}
\description{
Optimize XCMS parameters (detect PPM cutoff) using a binned EIC approach.
Fixed for parallel processing environment isolation.
}
